Updated Grabbing and Releasing object
Now if XR Grab Interactable has an attach transform object on it then it will be able to grab in the desired orientation.

Now Releasing Object will release it either 3f outwards or whatever collider you are looking at

Object position is set to right infront of user just for testing right now.
Also drop distance in set to 3f for testing can be changed for desired behavior later